requirements.txt
SQLAlchemy
graphene_sqlalchemy
Flask
Flask-GraphQL
pipenv
pytest
python-dotenv
PyMySQL
requests
